Brings the branch current with upstream (#814, #818, #819, #820, #821, #823,
#825). No changes of our own -- upstream's content only.

One conflict, in tests/test_AFC_hub.py, where upstream refactored the
_make_hub fixture to build through the real __init__/handle_connect; resolved
to upstream's side.

+++ extras/AFC_stepper.py
@@ -692,7 +692,7 @@ class AFCExtruderStepper(AFCLane):
         self.logger.debug(f"{self.name} adding endstop {key}:{name}:{pin}") # TODO:remove once fully tested on toolchanger
         self._endstops[key] = (mcu_endstop, name)
 
-    def do_homing_move(self, movepos: int, speed: int, accel: int, endstop_spec:str,
+    def do_homing_move(self, movepos: float, speed: float, accel: float, endstop_spec:str,
                        triggered=True, check_trigger=True, assist_active=True) -> tuple[bool, float]:
         """
         Perform's a homing move using the specified endstop, speed/accel and distance.
